Move lexer and parser into class
